🌈feat : selectBox 설명 추가
