+++ client/package.json
@@ -14,7 +14,11 @@
     "formatter": "prettier --write --cache \"src/**/*.{ts,tsx}\""
   },
   "lint-staged": {
-    "src/**/*.{js,ts,jsx,tsx}": ["npm run lint-fix", "npm run format", "git add"]
+    "src/**/*.{js,ts,jsx,tsx}": [
+      "npm run lint-fix",
+      "npm run format",
+      "git add"
+    ]
   },
   "husky": {
     "hooks": {
@@ -32,6 +36,7 @@
     "esbuild": "^0.19.2",
     "react": "^18.2.0",
     "react-router-dom": "^6.15.0",
+    "storybook-addon-react-router-v6": "^2.0.7",
     "styled-components": "^5.3.1",
     "styled-components.macro": "^1.0.0"
   },
