+++ client/src/App.tsx
@@ -1,12 +1,12 @@
 import { RouterProvider } from 'react-router-dom';
 import router from '@router/Router';
-import KirriProvider from '@router/KirriProvider';
+import KkliriProvider from '@router/KkliriProvider';
 
 function App() {
   return (
-    <KirriProvider>
+    <KkliriProvider>
       <RouterProvider router={router} />
-    </KirriProvider>
+    </KkliriProvider>
   );
 }
